+++ packages/cli/src/main.ts
@@ -6,10 +6,10 @@ import consola, { Consola } from "consola"
 import { initErrorMonitoring } from "./services/error-monitoring/implementation.js"
 import { open } from "./commands/open/index.js"
 import { telemetry } from "./services/telemetry/implementation.js"
-import { getGitRemotes } from "./utilities/getGitRemotes.js"
-import { parseOrigin } from "@inlang/telemetry"
 import fetchPolyfill from "node-fetch"
 import { lint } from "./commands/lint/index.js"
+import { coreUsedConfigEvent } from "@inlang/telemetry"
+import { getConfig } from "./utilities/getConfig.js"
 
 // --------------- INIT ---------------
 
@@ -25,7 +25,6 @@ initErrorMonitoring()
 ;(consola as unknown as Consola).wrapConsole()
 
 // --------------- CLI ---------------
-const gitOrigin = parseOrigin({ remotes: await getGitRemotes() })
 
 export const cli = new Command()
 	.name("inlang")
@@ -42,8 +41,6 @@ export const cli = new Command()
 			(arg, i) => !arg.startsWith("-") && !command.args[i - 1]?.startsWith("-"),
 		)
 		telemetry.capture({
-			distinctId: "unknown",
-			groups: { repository: gitOrigin },
 			event: `CLI command executed`,
 			properties: {
 				name: name.join(" "),
@@ -57,10 +54,17 @@ export const cli = new Command()
 // not using await to not block the CLI
 
 telemetry.capture({
-	distinctId: "unknown",
-	groups: { repository: gitOrigin },
 	event: "CLI started",
 	properties: {
 		version,
 	},
 })
+
+const [inlangConfig] = await getConfig()
+
+if (inlangConfig) {
+	telemetry.capture({
+		event: coreUsedConfigEvent.name,
+		properties: coreUsedConfigEvent.properties(inlangConfig),
+	})
+}