+++ packages/website/src/pages/editor/@host/@owner/@repository/components/SearchInput.tsx
@@ -33,7 +33,7 @@ export const SearchInput = (props: SearchInputProps) => {
 	)
 }
 
-const SearchIcon = () => {
+export const SearchIcon = () => {
 	return (
 		<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20">
 			<path